pyproject.toml
[tool.poetry]
name = "flex-api"
version = "0.1.0"
description = ""
authors = ["vjanz <valon.januzaj98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.12.2"
alembic = "^1.4.3"
pydantic = {extras = ["email"], version = "^1.6.1"}
psycopg2 = "^2.8.6"
fastapi-utils = "^0.2.1"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
mypy = "^0.790"
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.6.4"
autoflake = "^1.4"
sqlalchemy-stubs = "^0.3"

[tool.black]
line-length = 100
target_version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | build
  )/
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
